Move helper stuff out of istio_controller_test.go
